LayoutTests/http/tests/resourceLoadStatistics/clear-in-memory-and-persistent-store-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS Host classified as prevalent resource.
PASS Host is no longer a prevalent resource after the store was cleared.
